+++ jgiven-junit/src/test/java/com/tngtech/jgiven/junit/test/WhenTestStep.java
@@ -2,6 +2,8 @@ package com.tngtech.jgiven.junit.test;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import org.junit.Assume;
+
 import com.tngtech.jgiven.Stage;
 import com.tngtech.jgiven.annotation.ExpectedScenarioState;
 import com.tngtech.jgiven.annotation.ProvidedScenarioState;
@@ -28,5 +30,9 @@ public class WhenTestStep extends Stage<WhenTestStep> {
 
     public void some_assertion_fails() {
         assertThat( true ).isFalse();
+    }
+
+    public void some_assumption_fails() {
+        Assume.assumeTrue( false );
     }
 }